somthing

+++ Dotnet2026_4049_0328/DalFacade/DO/Sale.cs
@@ -9,7 +9,7 @@ public record Sale(
     int SaleId,
     int ProductId,
     int QuantityToSale,
-    int TotalPrice,
+    double TotalPrice,
     bool IsClube,
     DateTime StartSale,
     DateTime EndSale
